requirements.txt
asgiref==3.8.1
asttokens==3.0.0
coverage==7.9.1
decorator==5.2.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-stubs==5.2.1
django-stubs-ext==5.2.1
dotenv==0.9.9
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.0
freezegun==1.5.2
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
wcwidth==0.2.13
